Reduce NIN AF2 wait from 2 vana days to 1
This seems to have just been a mistake when porting the Ninja AF quests
to the new interaction framework. After trading the Oak Pole to
Leodarion, the quest checks to see if the current day _is_ or _comes
before_ the day the trade happened; if so, it plays the "come back
later" event. However, when recording the day the trade happened we are
actually recording the day _after_ the trade happened. This difference
causes the player to wait 2 game days instead of 1.